+++ packages/cli/typescript.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+var ethers_1 = require("ethers");
 function getType(param, flexible) {
     if (param.type === "address" || param.type === "string") {
         return "string";
@@ -105,7 +106,7 @@ function generate(contract, bytecode) {
     lines.push("    static ABI(): Array<string> {");
     lines.push("        return [");
     contract.interface.fragments.forEach(function (fragment) {
-        lines.push("            \"" + fragment.format(true) + "\",");
+        lines.push("            \"" + fragment.format(ethers_1.ethers.utils.FormatTypes.full) + "\",");
     });
     lines.push("        ];");
     lines.push("    }");